+++ ydb/services/sqs_topic/actor.cpp
@@ -1,0 +1,1 @@
+#include "actor.h"
